+++ src/main/cpp/blaze_util.h
@@ -42,17 +42,17 @@ std::string GetUserName();
 std::string MakeAbsolute(const std::string &path);
 
 // Replaces 'content' with contents of file 'filename'.
-// If `max_size` is positive, the method reads at most that many bytes; if it
-// is 0, the method reads the whole file.
+// If `max_size` is positive, the method reads at most that many bytes;
+// otherwise the method reads the whole file.
 // Returns false on error. Can be called from a signal handler.
 bool ReadFile(const std::string &filename, std::string *content,
-              size_t max_size = 0);
+              int max_size = 0);
 
 // Replaces 'content' with contents of file descriptor 'fd'.
 // If `max_size` is positive, the method reads at most that many bytes; if it
-// is 0, the method reads the whole file.
+// otherwise the method reads the whole file.
 // Returns false on error. Can be called from a signal handler.
-bool ReadFileDescriptor(int fd, std::string *content, size_t max_size = 0);
+bool ReadFileDescriptor(int fd, std::string *content, int max_size = 0);
 
 // Writes 'content' into file 'filename', and makes it executable.
 // Returns false on failure, sets errno.
